Fixing audio installation issues
On oIWD, the old audio decompression scripts were failing due to a bad path to oggdec. There's an installation fork between
IWDEE/IWD-in-BG2 and oIWD, and the former already uses HANDLE_AUDIO. HANDLE_AUDIO also works on the original, so I simply moved
the H_A to the general install instructions and deleted the old scripting.